LPS-124963 Move to helper package
